requirements.txt
et-xmlfile==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
openpyxl==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
